promote Unison.Codebase.CommandLine to Unison.CommandLine
